[NVIM] move vim-options and keymaps settings in the config dir
