+++ nvim/.config/nvim/init.lua
@@ -1,4 +1,3 @@
-require("vim-options")
+require("config.options")
 require("config.lazy")
-
-vim.keymap.set("n", "<leader>ru", require("smart_refs").smart_references, { desc = "Smart Find References" })
+require("config.keymaps")
