+++ nvim/.config/nvim/lua/config/keymaps.lua
@@ -1,0 +1,1 @@
+vim.keymap.set("n", "<leader>ru", require("smart_refs").smart_references, { desc = "Smart Find References" })
