+++ nvim/.config/nvim/lua/config/options.lua
